# Conflicts:
#	app/src/main/java/net/fpoly/dailymart/view/order/OrderActivity.kt
#	app/src/main/java/net/fpoly/dailymart/view/order/OrderViewModel.kt

+++ app/src/main/java/net/fpoly/dailymart/utils/PriceUnit.kt
@@ -2,7 +2,7 @@ package net.fpoly.dailymart.utils
 
 import java.text.DecimalFormat
 
-fun Double.toMoney(): String {
+fun Int.toMoney(): String {
     val df = DecimalFormat("###,###,###.## ")
     return df.format(this) + " vnđ"
 }
